Refactored sym tool
