+++ src/pins2lts-sym/Makefile.am
@@ -4,7 +4,25 @@ SYM_CPP_FLAGS = $(AM_CPPFLAGS) $(PROFILER_CPPFLAGS)
 SYM_LD_FLAGS = $(AM_LDFLAGS) $(PROFILER_LDFLAGS)
 SYM_C_FLAGS = $(AM_CFLAGS) $(PROFILER_CFLAGS)
 
-LIBZZZ = ../pins-lib/libpins.la
+
+noinst_LTLIBRARIES  = libpins2lts_sym.la
+
+# libpins2lts_sym.la
+libpins2lts_sym_la_SOURCES  = pins2lts-sym.c aux/options.h
+libpins2lts_sym_la_SOURCES  += alg/aux.c alg/aux.h
+libpins2lts_sym_la_SOURCES  += alg/bfs.c alg/bfs.h
+libpins2lts_sym_la_SOURCES  += alg/chain.c alg/chain.h
+libpins2lts_sym_la_SOURCES  += alg/pg.c alg/pg.h
+libpins2lts_sym_la_SOURCES  += alg/mu.c alg/mu.h
+libpins2lts_sym_la_SOURCES  += alg/reach.c alg/reach.h
+libpins2lts_sym_la_SOURCES  += alg/sat.c alg/sat.h
+libpins2lts_sym_la_SOURCES  += aux/output.c aux/output.h
+libpins2lts_sym_la_SOURCES  += aux/prop.c aux/prop.h
+libpins2lts_sym_la_CPPFLAGS = $(SYM_CPP_FLAGS)
+libpins2lts_sym_la_LDFLAGS  = $(SYM_LD_FLAGS)
+
+LIBZZZ  = libpins2lts_sym.la
+LIBZZZ += ../pins-lib/libpins.la
 LIBZZZ += ../vset-lib/libvset.la
 LIBZZZ += maxsum/libmaxsum.la
 LIBZZZ += ../spg-lib/libspg.la
@@ -54,7 +72,7 @@ bin_PROGRAMS += prob2lts-sym
 endif
 
 # etf2lts-sym
-etf2lts_sym_SOURCES   = options.c pins2lts-sym.c options.h
+etf2lts_sym_SOURCES   = aux/options.c
 etf2lts_sym_SOURCES  += ../pins-lib/modules/etf-pins.c ../pins-lib/modules/etf-pins.h
 etf2lts_sym_LDADD     = $(LIBZZZ)
 etf2lts_sym_CFLAGS    = $(SYM_C_FLAGS)
@@ -63,7 +81,7 @@ etf2lts_sym_LDFLAGS   = $(SYM_LDFLAGS)
 nodist_EXTRA_etf2lts_sym_SOURCES = automake-force-linker.cxx # req. by vset
 
 # pins2lts-sym
-pins2lts_sym_SOURCES   = options.c pins2lts-sym.c options.h
+pins2lts_sym_SOURCES   = aux/options.c
 pins2lts_sym_SOURCES  += ../pins-lib/modules/dlopen-pins.c ../pins-lib/modules/dlopen-pins.h
 pins2lts_sym_LDADD     = $(LIBZZZ)
 pins2lts_sym_CFLAGS    = $(SYM_C_FLAGS)
@@ -74,7 +92,7 @@ endif
 nodist_EXTRA_pins2lts_sym_SOURCES = automake-force-linker.cxx # req. by vset
 
 # lpo2lts-sym
-lpo2lts_sym_SOURCES   = options.c pins2lts-sym.c options.h
+lpo2lts_sym_SOURCES   = aux/options.c
 lpo2lts_sym_LDADD     = $(LIBZZZ) $(MCRL_LIBS)
 lpo2lts_sym_LDADD    += ../pins-lib/libmcrl-pins.la
 lpo2lts_sym_CFLAGS    = $(SYM_C_FLAGS)
@@ -83,7 +101,7 @@ lpo2lts_sym_LDFLAGS   = $(MCRL_LDFLAGS) $(SYM_LD_FLAGS)
 nodist_EXTRA_lpo2lts_sym_SOURCES = automake-force-linker.cxx # req. by mcrl
 
 # lps2lts-sym
-lps2lts_sym_SOURCES   = options.c pins2lts-sym.c options.h
+lps2lts_sym_SOURCES   = aux/options.c
 lps2lts_sym_LDADD     = $(LIBZZZ) 
 lps2lts_sym_LDADD    += ../pins-lib/libmcrl2-pins.la
 lps2lts_sym_CFLAGS    = $(SYM_C_FLAGS)
@@ -92,7 +110,7 @@ lps2lts_sym_LDFLAGS   = $(MCRL2_LDFLAGS) $(SYM_LD_FLAGS)
 nodist_EXTRA_lps2lts_sym_SOURCES = automake-force-linker.cxx # req. by mcrl2
 
 # mapa2lts-sym
-mapa2lts_sym_SOURCES   = options.c pins2lts-sym.c options.h
+mapa2lts_sym_SOURCES   = aux/options.c
 mapa2lts_sym_LDADD     = ../pins-lib/libmapa-pins.la
 mapa2lts_sym_LDADD    += $(LIBZZZ) -lscoop
 mapa2lts_sym_LDFLAGS   = -L${top_builddir}/scoop/src $(SYM_LD_FLAGS)
@@ -101,7 +119,7 @@ mapa2lts_sym_CPPFLAGS  = -DMAPA $(SYM_CPP_FLAGS)
 nodist_EXTRA_mapa2lts_sym_SOURCES = automake-force-linker.cxx # req. by vset
 
 # pbes2lts-sym
-pbes2lts_sym_SOURCES   = options.c pins2lts-sym.c options.h
+pbes2lts_sym_SOURCES   = aux/options.c
 pbes2lts_sym_LDADD     = $(LIBZZZ)
 pbes2lts_sym_LDADD    += ../pins-lib/libpbes-pins.la
 pbes2lts_sym_CFLAGS    = $(SYM_C_FLAGS)
@@ -110,7 +128,7 @@ pbes2lts_sym_LDFLAGS   = $(MCRL2_LDFLAGS) $(SYM_LD_FLAGS)
 nodist_EXTRA_pbes2lts_sym_SOURCES = automake-force-linker.cxx # req. by pbes
 
 # dve2lts-sym
-dve2lts_sym_SOURCES   = options.c pins2lts-sym.c options.h
+dve2lts_sym_SOURCES   = aux/options.c
 dve2lts_sym_SOURCES  += ../pins-lib/modules/dve-pins.c ../pins-lib/modules/dve-pins.h
 dve2lts_sym_LDADD     = $(LIBZZZ) $(DVE_LIBS)
 dve2lts_sym_CFLAGS    = $(SYM_C_FLAGS)
@@ -119,7 +137,7 @@ dve2lts_sym_LDFLAGS   = $(DIVINE_LDFLAGS) $(SYM_LD_FLAGS)
 nodist_EXTRA_dve2lts_sym_SOURCES = automake-force-linker.cxx # req. by vset
 
 # prom2lts-sym
-prom2lts_sym_SOURCES  = options.c pins2lts-sym.c options.h
+prom2lts_sym_SOURCES  = aux/options.c
 prom2lts_sym_SOURCES += ../pins-lib/modules/prom-pins.c ../pins-lib/modules/prom-pins.h
 prom2lts_sym_LDADD    = $(LIBZZZ) $(SPINS_LIBS)
 prom2lts_sym_CFLAGS    = $(SYM_C_FLAGS)
@@ -128,7 +146,7 @@ prom2lts_sym_LDFLAGS  = $(SPINS_LDFLAGS) $(SYM_LD_FLAGS)
 nodist_EXTRA_prom2lts_sym_SOURCES = automake-force-linker.cxx # req. by vset
 
 # pnml2lts-sym
-pnml2lts_sym_SOURCES   = options.c pins2lts-sym.c options.h
+pnml2lts_sym_SOURCES   = aux/options.c
 pnml2lts_sym_SOURCES  += ../pins-lib/modules/pnml-pins.c ../pins-lib/modules/pnml-pins.h
 pnml2lts_sym_LDADD     = $(PNML_LIBS) $(LIBZZZ) ../andl-lib/libandl.la
 pnml2lts_sym_CPPFLAGS  = -DPNML $(SYM_CPP_FLAGS)
@@ -137,7 +155,7 @@ pnml2lts_sym_CFLAGS    = $(SYM_C_FLAGS) $(PNML_CFLAGS)
 nodist_EXTRA_pnml2lts_sym_SOURCES = automake-force-linker.cxx # req. by vset
 
 # prob2lts-sym
-prob2lts_sym_SOURCES   = options.c pins2lts-sym.c options.h
+prob2lts_sym_SOURCES   = aux/options.c
 prob2lts_sym_SOURCES  += ../pins-lib/modules/prob-pins.c ../pins-lib/modules/prob-pins.h
 prob2lts_sym_LDADD     = $(PROB_LIB_LIBS) $(LIBZZZ)
 prob2lts_sym_CFLAGS    = $(SYM_CFLAGS) $(PROB_CFLAGS)
