+++ src/pins2lts-sym/alg/sat.c
@@ -1,0 +1,297 @@
+#include <hre/config.h>
+
+#ifdef __APPLE__
+#define _DARWIN_C_SOURCE
+#endif
+
+#include <float.h>
+#include <limits.h>
+#include <stdbool.h>
+#include <stdint.h>
+#include <stdlib.h>
+#include <string.h>
+#include <strings.h>
+
+#include <dm/dm.h>
+#include <hre/stringindex.h>
+#include <hre/user.h>
+#include <lts-io/user.h>
+#include <pins-lib/pins.h>
+#include <pins-lib/pins-impl.h>
+#include <pins-lib/pins-util.h>
+#include <pins-lib/pins2pins-group.h>
+#include <pins2lts-sym/alg/aux.h>
+#include <pins2lts-sym/alg/sat.h>
+#include <pins2lts-sym/aux/output.h>
+#include <pins2lts-sym/aux/prop.h>
+#include <ltsmin-lib/ltsmin-standard.h>
+#include <ltsmin-lib/ltsmin-syntax.h>
+#include <ltsmin-lib/ltsmin-tl.h>
+#include <mc-lib/atomics.h>
+#include <mc-lib/bitvector-ll.h>
+#include <util-lib/bitset.h>
+#include <util-lib/dfs-stack.h>
+#include <util-lib/util.h>
+
+
+#ifdef HAVE_SYLVAN
+#include <sylvan.h>
+#else
+#define LACE_ME
+#define lace_suspend()
+#define lace_resume()
+#endif
+
+void
+reach_sat_fix(reach_proc_t reach_proc, vset_t visited,
+              bitvector_t *reach_groups, long *eg_count, long *next_count, long *guard_count)
+{
+    (void) reach_proc;
+    (void) guard_count;
+
+    if (PINS_USE_GUARDS)
+        Abort("guard-splitting not supported with saturation=sat-fix");
+
+    int level = 0;
+    vset_t old_vis = vset_create(domain, -1, NULL);
+    vset_t deadlocks = dlk_detect?vset_create(domain, -1, NULL):NULL;
+    vset_t dlk_temp = dlk_detect?vset_create(domain, -1, NULL):NULL;
+
+    LACE_ME;
+    while (!vset_equal(visited, old_vis)) {
+        if (trc_output != NULL) save_level(visited);
+        vset_copy(old_vis, visited);
+        stats_and_progress_report(NULL, visited, level);
+        level++;
+        for(int i = 0; i < nGrps; i++){
+            if (!bitvector_is_set(reach_groups, i)) continue;
+            expand_group_next(i, visited);
+            reach_chain_stop();
+            (*eg_count)++;
+        }
+        if (dlk_detect) vset_copy(deadlocks, visited);
+        vset_least_fixpoint(visited, visited, group_next, nGrps);
+        (*next_count)++;
+        check_invariants(visited, level);
+        if (dlk_detect) {
+            for (int i = 0; i < nGrps; i++) {
+                vset_prev(dlk_temp, visited, group_next[i],deadlocks);
+                reduce(i, dlk_temp);
+                vset_minus(deadlocks, dlk_temp);
+                vset_clear(dlk_temp);
+            }
+            deadlock_check(deadlocks, reach_groups);
+        }
+        vset_reorder(domain);
+    }
+
+    vset_destroy(old_vis);
+    if (dlk_detect) {
+        vset_destroy(deadlocks);
+        vset_destroy(dlk_temp);
+    }
+}
+
+static void
+initialize_levels(bitvector_t *groups, int *empty_groups, int *back,
+                      bitvector_t *reach_groups)
+{
+    int level[nGrps];
+
+    // groups: i = 0 .. nGrps - 1
+    // vars  : j = 0 .. N - 1
+
+    // level[i] = first '+' in row (highest in BDD) of group i
+    // recast 0 .. N - 1 down to equal groups 0 .. (N - 1) / sat_granularity
+    for (int i = 0; i < nGrps; i++) {
+        level[i] = -1;
+
+        for (int j = 0; j < N; j++) {
+            if (dm_is_set(GBgetDMInfo(model), i, j)) {
+                level[i] = (N - j - 1) / sat_granularity;
+                break;
+            }
+        }
+
+        if (level[i] == -1)
+            level[i] = 0;
+    }
+
+    for (int i = 0; i < nGrps; i++)
+        bitvector_set(&groups[level[i]], i);
+
+    // Limit the bit vectors to the groups we are interested in and establish
+    // which saturation levels are not used.
+    for (int k = 0; k < max_sat_levels; k++) {
+        bitvector_intersect(&groups[k], reach_groups);
+        empty_groups[k] = bitvector_is_empty(&groups[k]);
+    }
+
+    if (back == NULL)
+        return;
+
+    // back[k] = last + in any group of level k
+    bitvector_t level_matrix[max_sat_levels];
+
+    for (int k = 0; k < max_sat_levels; k++) {
+        bitvector_create(&level_matrix[k], N);
+        back[k] = max_sat_levels;
+    }
+
+    for (int i = 0; i < nGrps; i++) {
+        dm_row_union(&level_matrix[level[i]], GBgetDMInfo(model), i);
+    }
+
+    for (int k = 0; k < max_sat_levels; k++) {
+        for (int j = 0; j < k; j++) {
+            bitvector_t temp;
+            int empty;
+
+            bitvector_copy(&temp, &level_matrix[j]);
+            bitvector_intersect(&temp, &level_matrix[k]);
+            empty = bitvector_is_empty(&temp);
+            bitvector_free(&temp);
+
+            if (!empty)
+                if (j < back[k]) back[k] = j;
+        }
+
+        if (back[k] == max_sat_levels && !bitvector_is_empty(&level_matrix[k]))
+            back[k] = k + 1;
+    }
+
+    for (int k = 0; k < max_sat_levels; k++)
+        bitvector_free(&level_matrix[k]);
+
+}
+
+void
+reach_sat_like(reach_proc_t reach_proc, vset_t visited,
+               bitvector_t *reach_groups, long *eg_count, long *next_count, long *guard_count)
+{
+    bitvector_t groups[max_sat_levels];
+    int empty_groups[max_sat_levels];
+    int back[max_sat_levels];
+    int k = 0;
+    int last = -1;
+    vset_t old_vis = vset_create(domain, -1, NULL);
+    vset_t prev_vis[nGrps];
+
+    for (int k = 0; k < max_sat_levels; k++)
+        bitvector_create(&groups[k], nGrps);
+
+    initialize_levels(groups, empty_groups, back, reach_groups);
+
+    for (int i = 0; i < max_sat_levels; i++)
+        prev_vis[i] = save_sat_levels?vset_create(domain, -1, NULL):NULL;
+
+    while (k < max_sat_levels) {
+        if (k == last || empty_groups[k]) {
+            k++;
+            continue;
+        }
+
+        Warning(infoLong, "Saturating level: %d", k);
+        vset_copy(old_vis, visited);
+        reach_proc(visited, prev_vis[k], &groups[k], eg_count, next_count,guard_count);
+        check_invariants(visited, -1);
+        if (save_sat_levels) vset_copy(prev_vis[k], visited);
+        if (vset_equal(old_vis, visited))
+            k++;
+        else {
+            last = k;
+            k = back[k];
+        }
+    }
+
+    for (int k = 0; k < max_sat_levels; k++)
+        bitvector_free(&groups[k]);
+
+    vset_destroy(old_vis);
+    if (save_sat_levels)
+        for (int i = 0; i < max_sat_levels; i++) vset_destroy(prev_vis[i]);
+}
+
+void
+reach_sat_loop(reach_proc_t reach_proc, vset_t visited,
+               bitvector_t *reach_groups, long *eg_count, long *next_count, long *guard_count)
+{
+    bitvector_t groups[max_sat_levels];
+    int empty_groups[max_sat_levels];
+    vset_t old_vis = vset_create(domain, -1, NULL);
+    vset_t prev_vis[nGrps];
+
+    for (int k = 0; k < max_sat_levels; k++)
+        bitvector_create(&groups[k], nGrps);
+
+    initialize_levels(groups, empty_groups, NULL, reach_groups);
+
+    for (int i = 0; i < max_sat_levels; i++)
+        prev_vis[i] = save_sat_levels?vset_create(domain, -1, NULL):NULL;
+
+    while (!vset_equal(old_vis, visited)) {
+        vset_copy(old_vis, visited);
+        for (int k = 0; k < max_sat_levels; k++) {
+            if (empty_groups[k]) continue;
+            Warning(infoLong, "Saturating level: %d", k);
+            reach_proc(visited, prev_vis[k], &groups[k], eg_count, next_count,guard_count);
+            check_invariants(visited, -1);
+            if (save_sat_levels) vset_copy(prev_vis[k], visited);
+        }
+    }
+
+    for (int k = 0; k < max_sat_levels; k++)
+        bitvector_free(&groups[k]);
+
+    vset_destroy(old_vis);
+    if (save_sat_levels)
+        for (int i = 0; i < max_sat_levels; i++) vset_destroy(prev_vis[i]);
+}
+
+void
+reach_sat(reach_proc_t reach_proc, vset_t visited,
+          bitvector_t *reach_groups, long *eg_count, long *next_count, long *guard_count)
+{
+    (void) reach_proc;
+    (void) next_count;
+    (void) guard_count;
+
+    if (PINS_USE_GUARDS)
+        Abort("guard-splitting not supported with saturation=sat");
+
+    if (act_detect != NULL && trc_output != NULL)
+        Abort("Action detection with trace generation not supported");
+
+    for (int i = 0; i < nGrps; i++) {
+        if (bitvector_is_set(reach_groups, i)) {
+            struct expand_info *ctx = RTmalloc(sizeof(struct expand_info));
+            ctx->group = i;
+            ctx->group_explored = group_explored[i];
+            ctx->eg_count = eg_count;
+
+            vrel_set_expand(group_next[i], expand_group_next_projected, ctx);
+        }
+    }
+
+    if (trc_output != NULL) save_level(visited);
+    stats_and_progress_report(NULL, visited, 0);
+    vset_least_fixpoint(visited, visited, group_next, nGrps);
+    stats_and_progress_report(NULL, visited, 1);
+
+    check_invariants(visited, -1);
+
+    if (dlk_detect) {
+        vset_t deadlocks = vset_create(domain, -1, NULL);
+        vset_t dlk_temp = vset_create(domain, -1, NULL);
+        vset_copy(deadlocks, visited);
+        for (int i = 0; i < nGrps; i++) {
+            vset_prev(dlk_temp, visited, group_next[i],deadlocks);
+            reduce(i, dlk_temp);
+            vset_minus(deadlocks, dlk_temp);
+            vset_clear(dlk_temp);
+        }
+        deadlock_check(deadlocks, reach_groups);
+        vset_destroy(deadlocks);
+        vset_destroy(dlk_temp);
+    }
+}
